+++ app/admin/payments/page.tsx
@@ -1,14 +1,12 @@
-import { getAllTransactions, getTenantOptions, getApartmentsWithActiveLease } from '@/lib/adminData'
+import { getTenantOptions, getApartmentsWithActiveLease } from '@/lib/adminData'
 import { getLinxoTransactions } from '@/lib/linxoImport'
-import PaymentsClient from '@/components/admin/PaymentsClient'
 import LinxoTable from '@/components/admin/LinxoTable'
 import QuittancesGenerator from '@/components/admin/QuittancesGenerator'
 
 export const dynamic = 'force-dynamic'
 
 export default async function AdminPaymentsPage() {
-  const [transactions, linxoRows, tenantOptions, apartments] = await Promise.all([
-    getAllTransactions(),
+  const [linxoRows, tenantOptions, apartments] = await Promise.all([
     getLinxoTransactions().catch(() => []),
     getTenantOptions().catch(() => []),
     getApartmentsWithActiveLease().catch(() => []),
@@ -22,7 +20,6 @@ export default async function AdminPaymentsPage() {
         initialRows={linxoRows}
         tenantOptions={tenantOptions}
       />
-      <PaymentsClient transactions={transactions} />
     </div>
   )
 }